Extract calendar feed endpoint
